refactor: move api classes to their own modules
