+++ altmetric/__init__.py
@@ -1,3 +1,1 @@
-from .explorer import api
 
-__all__ = ['api']
