+++ altmetric/explorer/api/__init__.py
@@ -1,0 +1,3 @@
+from .client import Client
+
+__all__ = ['Client']
