+++ altmetric/explorer/api/test_digest.py
@@ -1,8 +1,9 @@
 import pytest
 
-from . import api
+from .client import digest
 
 API_SECRET = 'my_secret_key'
+
 
 @pytest.mark.parametrize("message,expected", [
     ('', '4b4f493acb45332879e4812a98473fc98209fee6'),
@@ -10,4 +11,4 @@ API_SECRET = 'my_secret_key'
     ('foo', '155ce81b88766ccd779d515af9ab6081586f077f'),
 ])
 def test_calculating_a_digest(message, expected):
-    assert api.digest(API_SECRET, message) == expected
+    assert digest(API_SECRET, message) == expected
